# Better encyption system and enhanced custom table widget
## Encryption:
Now, when hashing a password, a unique salt will be added, and will add more security to the encryption system.

## CustomTable:
The CustomTable widget now can, when the column is of type Number, add cells between them.
This mean that if the column A, B, and C (1, 2 and 3) are of type number, the cell at (C1), so the first cell of the column C can contain a calculation, that will start with the = sign.

if the cell at (A1) equals 1 and the cell at (B1) equals 2, the input =(A1)+(B1) will output 3.

I also added a way to extend cells data.

This means that you can now drag a little blue point on a cell, that will let you select mutiples cells from a same column, and extend the value from the first cell to the others.

If the column is of type Number, and the value of the first selected cell is a calculation like seen before, the rows positions will be incremented or decremented for the number of rows.