Enhanced encryption, and finished CustomTableCells
Switched from sha256 to sha512 for password hashing.

Finished all custom CustomTableCell by implementing the selection cell, as well as changing the date cell to let users add an end date.

Added a DefaultAppColors, which contains colors that can be used by users to customize an option in a selection cell for example.