Started to create a node system for a future plugin system.
The node tree will run asynchronously, and return a value via the run method, if an IOutpout node is given. Otherwise returns null.

Added a folder in assets named code_templates that will contain templates of code for repetitive code snippets.

Updated the createUniqueID method for more precision, as repeted, close in time calls, where likely going to return the same id.

Started a new page named PluginEditorView.
This page will be where users will be able to create plugins for minddy

Added a real read only mode for articles, so if the read only mode is enabled on an article, for example in the default articles on the homepage, the content will not be editable or deletable.