Started to create a PluginEditorView, by creating a node system.
For now, this view contains a NodeEditorBottomSheet element, that contains all the nodes. This bottom sheet can be closed to give more space, for what will be a UI editor.

Nodes can be connected, and disconnected, connection that is shown by a 'cable' between them.

Each node can connect to others via a 'NodePort' that contains every informations that is needed like the port index, the node itself, and some functions that allow the port to control connections between nodes.

All of those functions are from the NodeEditorBottomSheetController, that controls everything.

The whole thing is saved as json, and stored at 'minddy/shared/plugins/'

Still a lot of work to get the code cleaner, and more usable.

I now have to work on node selection, the editor itself, like selecting a few nodes and moving them, deleting them, and things like that.