Added node selection logic, and basic undo/redo logic
-Major changes

Added the side panel to the NodeEditorBottomSheet. This side panel will contain node trees and variables.

In this side panel, i added the node tree selector, which will allow users to change the selected node tree.

Clicking a node deselect every other one, except if the shift key is pressed.

Clicking the grid behind the nodes will deselect every node

If you drag the node board while pressing shift, and a box selector will let you select nodes that are overapped by this box.

A selected node takes the color of the secondary color, of the actual theme.

If multiples nodes are selected, the last one in the list takes the secondary color, the other ones takes the DefaultAppColors.yellow color.

The nodes connections that are directly linked to a selected node, will also appear as selected, with the same secondary color.

If multiples nodes are selected at the same time, and you start dragging one of them, all of the selected nodes will follow.

Also, when using a trackpad, moving the trackpad while hovering a node will no longer move it. Same for nodes ports.

Node ports positions are now automaticly calculated at the build time. This mean that the construction of the NodeConnections is a bit delayed.

Added a basic undo/redo logic to the NodeBottomSheetEditor.

-Minor changes

Added a functionName parameter to the function _handleError, and _handleEncryptionError, for easier debbuging + Removed commented out testing code.

Added a message to the Pomodoro tool, which now shows a snackbar indicating "Work time" or "Break time".

Added a redundant color error message to the CustomTableSelectionCellAddOptionsSubMenuController. Now two option can't have the same color.