Added a copy/paste system to the nodes, a delete system, and the variable system.
- Major changes

Users can now press cmd(ctrl)+backspace to delete all the selected nodes.

Users can also press cmd(ctrl)+c to copy the selected nodes, cmd(ctrl)+v to paste them, or just cmd(ctrl)+d to duplicate, which does the copy and paste at the same time.

Added a new class named NodeTreeVariable, which contains the variable data like the name, id, type and value.
The variables are accessible by the side panel on the NodeBottomSheetEditor, where users will be able to rename variables, as well as select the type.

I also added a toolbar on the bottom of the NodeBottomSheetEditor, that allow easy access to the functions like copy, paste, duplicate, delete, or even undo and redo.

-Minor changes

Added a border to every sub menu, and some buttons to make the ui more persistent.

Removed the widget, int and float types in NodeDataType, and added a number type.

Created a shortcuts folder containing a shortcuts_activator.dart file, which will contain every shortcuts activator and intents for the app.