Added a better state handling in the NodeEditorBottomSheet, and added a menu to add nodes
- Major changes

Now, users can use the NodeEditorNewNodeSubMenu to add nodes to the NodeEditorBottomSheet. Featuring a search function, node sorting by category (Math, logic etc...) and a detail view where inputs and outputs types are shown.

Now, NodeEditorState are being saved for each NodeEditorBottomSheetView. This means that each view as it's own NodeEditorState, and selected nodes list.

The copied nodes list is now shared across every NodeEditorBottomSheetView, and is no longer specific to a NodeEditorState.

- Minor changes

Changed the borders on sub menus and buttons to be brighter in light mode.

INodeWidget now required a class named NodeWidgetFunctions instead of all of the functions. This allow simpler NodesWidgets declarations.

Enhanced article exporter to create copies of images, instead of using the same as the original article.