Simplified the way to create INodesWidgets.
- Major changes

Created a NodeWidgetInput, and NodeWidgetOutput widgets, that will allow easy nodes widgets creations.

Also created a NodeWidgetDropdown widget which allow easier creations of dropdowns, and a few other layers of abstraction like NodeWidgetBody and NodeWidgetInformations, that helps to create easily nodes widgets.

Added a node_widget_template.dart file in the assets/code_templates folder. It features a basic layout for nodes widgets.

- Minor Changes

Added a CustomCheckbox widget for more coherence over the app design.

Modified the NodeConnectionPainter and CurvedLinePainter to add a tilt when a node output is further than the input, which was causing the connection to be drawn behind the widget.

Changed the CustomSnackbar to use a blurred background instead of the theme.surface color.

Added a proper loading screen.

Added a AnimatedRouteBuilder class that helps to keep the same page transition animation across all platforms.

Udpated the NodeFunctions class to incorporate functions that ensure correct input changes.