New logo, new nodes widgets, and a new notification system
 - Major changes

 New ComparisonNodeWidget, BooleanNodeWidget

 New notification system, that uses a StreamBuilder at the lowest level of the Flutter node tree, to ensure simplicity.
A new class called NotificationsHandler will create a stream with all the notifications, stacked. -> It doesn't requires BuildContext like the old system.

- Minor changes

Updated all of the code where showBottomSnackbar was used, and replaced with the new NotificationHandler.addNotification.

Changed the isNodeBefore method in the NodeTree class to be isInputNodeBefore, to be more explicit.