Added time selection to CustomInputDatePicker, created a new CustomSelectionMenu.
 - Major changes

New switch option 'include hour' to the CustomDatePicker, that allows users to select an hour for a date.

Introducing a new CustomSelectionMenu widget that is closer to the Figma design, and includes search, sub-menus, cleaner design and more.

Created a new NodeTreeVariablesManager, that will help node trees to share the same variables across a single run.

  - Minor changes

Changed the date save format to use the iso8601 format.

Removed the startEnd option for the CustomDatePicker, replaced by the range option

Updated every dropdown menu of the app, to use the new CustomSelectionMenu.

Removed the node_types_interfaces.dart file, and placed the classes inside into single and different files for easier debugging.

Deleted the CustomPopupMenuButton, and modified the CustomDropdownButton to use the new CustomSelectionMenu