New SetVariable and GetVariable nodes, with a new variable handling system.
- Major changes

New SetVariableNodeWidget, that will help to define variables values with nodes.

New GetVariableNodeWidget that helps to get a variable value.

New RandomNumber node that takes a min and max as inputs.

Added the node description when hovering for one second a node widget title.

 - Minor changes

 Added search bars to the side panel boxes in the NodeEditorBottomSheet. Appears if the list of elements lenght is greater than or equal 4

Translated all the nodes categories and descriptions in the NewNodeSubMenu.