TextOperationNode, Private and colored projects, redesigned projects cards, new custom color picker and more...
 - Major changes

New TextOperationNode, that allows users to perform operations on texts in the NodeEditor.

Projects can now be set to 'private', which will require the user to enter the password to  get access to the project page.

Projects now have colors, to be tell apart more easily.

Projects are now sorted on the home page using the last change date.

New ProjectCard's and HomeViewArticleCards design with 'floating' titles, and hover effects.

New CustomColorPicker widget that allow users to select colors more precisely. Including a HEX and RGB code translator.

New ConversionNode that helps to convert data and types.

Because of the ConversionNode, type verifications when connecting two nodes is more strict. The types of input and output has to be the same for the connection to happen.

 - Minor changes

Changed the random number node to start with a min value of 0 and a max value of 1.

Removed the 'welcome snackbar' when unlocking with the password.

The articles content scrollbar now really reflects the article length.

Pomodoro tool will now show a snackbar on screen to annonce next rounds, even if the tool view is closed.