Uncutted NodePortWidget hitbox, new CustomSelectionMenu tooltips, Calendar system...
 - Major changes

NodePortWidget hitbox is no longer cut by the edge of the NodeWidgetBody widget.

CustomSelectionMenuItem now have tootip and tooltipTitle fields, allowing easier understanding of what the option is.

New Calendar and CalendarEvent classes, defining the base of what is a calendar in minddy.

New AppCalendar class, allowing easy access to common actions like creating, deleting or modifying a calendar or an event.

Enhcanced the CalendarButton, which can now be opened to see the next 10 events, and the next due dates.

CalendarEvents can have recurrences that can be individually modified or deleted.

CalendarEvents can be presented as due dates, to keep track of tasks within the calendar.

Uncompleted due dates are shown with red borders if the date is passed.

 - Minor changes

Reduced the NotificationWidget hitbox to the widget hitself, instead of taking the whole screen.

Modified the initializeApp Function to be easier to read and modify by working step by step.

Translated the calendar button part in all 3 languages.

Added a UiUpdater class that will be used to update ui using AnimatedBuilder, instead of creating a specific new Updater for each case. I will proceed in next commits to replace all of those updaters by the UiUpdater class.