Fixed infinite calendar files creation, search events, new time picker, finished event recurrence system and more...
 - Major changes

Fixed the infinite calendar files creation by encapsulating the CalendarButton on the HomeView in a FutureBuilder waiting for the app to be unlocked.

Changed the time picker to an easier caroussel solution, making date picking possible without touching the keyboard.

Added a search tab in the CalendarBottomMenu, which allow users to search for an event, click on it, and see it on the week view.

Finished the calendar event recurence system, introducing custom recurrence creation.

 - Minor changes

Changed the side of the CalendarEventWeekPreview to a pill instead of a border.

Changed the time indicator in the CalendarWeekViewPanel to appear only if the selected week contains today. It also is thicker on the date of today for easier reading.

CalendarBottomMenu now stays open even after a new week is selected.

Fixed the bug where the scroll in the NewProjectSubMenu was not fluid.

Added a hover effect on calednar button's close and open week view buttons.