New AppConfigValues class, Stored NodeTreeVariables values, PluginUiComponents
 - Major changes

Introducing the AppConfigValues class, that will make config data changes and getting easier. This also means that a lot of async code has been changed to sync code.

NodeTreeVariables values are now stored with the variable's Json, making the values recoverable, and introducing a storage system to Plugins.

Introducing the PluginEditorUiView, allowing users to create ui elements.

Ui elements are managable via the side panel, that is on the right side of the PluginEditorUiView, allowing users to change the component's properties, and layers.

For now, available conponents are Container, Column and row.

 - Minor changes

Added a (kinda) null safety to GetVariableNode to ensure proper working.

NodeEditorBottomSheet now saves the open/close state of the sheet.